wishlist loading states on product list page done for now

+++ src/components/ProductList/ProductListItem.js
@@ -3,39 +3,48 @@ import Card from "../UI/Card/Card";
 import styles from "./ProductListItem.module.css";
 import toDollars from "../../utilities/toDollars";
 import Icon, { ICON_TYPE } from "../UI/Icon/Icon";
-import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import {
   removeFromWishlist,
   addToWishlist,
 } from "../../redux/actions/wishlist-actions";
 import { Link } from "react-router-dom";
+import Spinner from "../UI/Spinner/Spinner";
 
 const ProductListItem = ({ product }) => {
-  const wishlist = useSelector((state) => state.wishlist.products);
-  const navigate = useNavigate();
+  const {
+    products: wishlistProducts,
+    isLoading: wishlistIsLoading,
+    loadingProduct: wishlistIsLoadingProduct,
+  } = useSelector((state) => state.wishlist);
   const dispatch = useDispatch();
   const [inWishlist, setInWishlist] = useState(false);
 
   useEffect(() => {
-    const foundProduct = wishlist.find((p) => p.id === product.id);
+    const foundProduct = wishlistProducts.find((p) => p.id === product.id);
     setInWishlist(!!foundProduct);
-  }, [product, wishlist]);
+  }, [product, wishlistProducts]);
 
-  const productClickHandler = () => {
-    navigate(`/product/${product.id}`);
-  };
+  const productLoadingInWishlist = wishlistIsLoadingProduct === product.id;
 
   const toggleWishlistHandler = (e) => {
     e.preventDefault();
-    inWishlist
-      ? dispatch(removeFromWishlist(product.id))
-      : dispatch(addToWishlist(product));
+    if (!wishlistIsLoading && !productLoadingInWishlist) {
+      inWishlist
+        ? dispatch(removeFromWishlist(product.id))
+        : dispatch(addToWishlist(product));
+    }
   };
 
   const wishListIcon = inWishlist
     ? ICON_TYPE.HEART_FULL
     : ICON_TYPE.HEART_EMPTY;
+
+  const wishlistBtnClasses =
+    styles["wishlist-toggle"] +
+    (wishlistIsLoading || productLoadingInWishlist
+      ? ` ${styles["wishlist-toggle--read-only"]}`
+      : "");
 
   return (
     <li className={styles["grid-flex"]}>
@@ -44,10 +53,17 @@ const ProductListItem = ({ product }) => {
           <div className={styles["image-wrapper"]}>
             <button
               title="Add to wishlist"
-              className={styles["wishlist-toggle"]}
+              className={wishlistBtnClasses}
               onClick={toggleWishlistHandler}
             >
-              <Icon icon={wishListIcon} />
+              {productLoadingInWishlist ? (
+                <Spinner
+                  className={styles["wishlist-spinner"]}
+                  spinnerWidth={0.15}
+                />
+              ) : (
+                <Icon icon={wishListIcon} />
+              )}
             </button>
             <img
               className={styles.image}
